Separate out tag scanning test
