+++ core/src/test/scala/akka/persistence/cassandra/journal/RecoveryLoadSpec.scala
@@ -6,22 +6,19 @@ package akka.persistence.cassandra.journal
 
 import akka.actor._
 import akka.persistence.{ PersistentActor, RecoveryCompleted, SnapshotOffer }
-import akka.persistence.cassandra.{ CassandraLifecycle, CassandraSpec, TestTaggingActor }
+import akka.persistence.cassandra.{ CassandraLifecycle, CassandraSpec }
 import akka.persistence.journal.Tagged
 import com.typesafe.config.ConfigFactory
-import org.scalatest.time.{ Milliseconds, Seconds, Span }
 
 import scala.concurrent.duration._
 
 object RecoveryLoadSpec {
   val config = ConfigFactory.parseString(s"""
       akka.loglevel = INFO
-      cassandra-journal.keyspace=RecoveryLoadSpec
       cassandra-journal.events-by-tag.enabled = on
       cassandra-journal.events-by-tag.scanning-flush-interval = 2s
       cassandra-journal.replay-filter.mode = off
       cassandra-journal.log-queries = off
-      cassandra-snapshot-store.keyspace=RecoveryLoadSpecSnapshot
       cassandra-snapshot-store.log-queries = off
     """).withFallback(CassandraLifecycle.config)
 
@@ -142,30 +139,6 @@ class RecoveryLoadSpec extends CassandraSpec(RecoveryLoadSpec.config) {
         println(s"iteration #$n")
         printMetrics(metrics)
         system.stop(p2)
-      }
-    }
-
-    "complete writes to tag scanning for many persistent actors" in {
-
-      val nrActors = 25
-      (0 until nrActors).foreach { i =>
-        val ref = system.actorOf(TestTaggingActor.props(s"$i"))
-        ref ! "msg"
-      }
-
-      awaitAssert {
-        import scala.collection.JavaConverters._
-        val expected = (0 until nrActors).map(n => (s"$n".toInt, 1L)).toList
-        val scanning = journalSession
-          .execute("select * from tag_scanning")
-          .all()
-          .asScala
-          .toList
-          .map(row => (row.getString("persistence_id"), row.getLong("sequence_nr")))
-          .filterNot(_._1 == "persistenceInit")
-          .map { case (pid, seqNr) => (pid.toInt, seqNr) } // sorting by pid makes the failure message easy to interpret
-          .sortBy(_._1)
-        scanning shouldEqual expected
       }
     }
   }
